Use temporary directory as user data directory for electron plugin playwright test
