+++ sbt-scalajs-esbuild-electron/src/sbt-test/sbt-scalajs-esbuild-electron/electron-project/app/src/main/scala/example/facade/node/Fs.scala
@@ -1,0 +1,10 @@
+package example.facade.node
+
+import scala.scalajs.js
+import scala.scalajs.js.annotation.JSImport
+
+object Fs {
+  @js.native
+  @JSImport("fs", "mkdtempSync")
+  def mkdtempSync(prefix: String): String = js.native
+}
