+++ sbt-scalajs-esbuild-electron/src/sbt-test/sbt-scalajs-esbuild-electron/electron-project/test
@@ -1,7 +1,7 @@
 $ absent app/esbuild/package-lock.json
 > clean
 > set app/Compile/scalaJSStage := org.scalajs.sbtplugin.Stage.FastOpt
-> integration-test/test
+> test
 > app/fullLinkJS/esbuildServeStart
 # TODO figure out how to test if dev mode works
 > app/fullLinkJS/esbuildServeStop
@@ -10,7 +10,7 @@ $ exists app/esbuild/package-lock.json
 $ delete app/esbuild/package-lock.json
 > clean
 > set app/Compile/scalaJSStage := org.scalajs.sbtplugin.Stage.FullOpt
-> integration-test/test
+> test
 > app/fullLinkJS/esbuildServeStart
 # TODO figure out how to test if dev mode works
 > app/fullLinkJS/esbuildServeStop
